backend/requirements.txt
python-docx==0.8.11
docx2pdf==0.1.8
Flask==2.3.3
flask-cors==4.0.0
